+++ src/App.tsx
@@ -5,7 +5,6 @@ import Footer from './components/footer/footer';
 import Popup from './components/popup/popup';
 
 function App() {
-//  alert('Привет. шахматы в процессе производства ещё. прошу не проверять работу пару дней, хочу хоть какой то функционал доделать. спасибо!');
   return (
     <>
       <Popup />
